feat: 타입 변경
int 타입으로 파라미터를 받아도 double 타입으로 계산되도록 수정

+++ src/lv3/App.java
@@ -20,7 +20,7 @@ public class App {
         OperatorType operatorType = OperatorType.getType(operator);
 
         // 테스트
-        int result = calculator.calculate(firstNumber, secondNumber, operatorType);
+        double result = calculator.calculate(firstNumber, secondNumber, operatorType);
         System.out.println("result = " + result);
     }
 }
